A LOT of work here. See description.
Added classes:
- Scene (representation of a level / menu / what ever you want to
render as a “scene”).
- SceneManager (loads scenes, change between them…).
- ScenesIdentifiers (enumeration of scene ID).
- SceneGraphNode (representation of a node of the scene graph).
- Entity (inherits from SceneGraphNode. Abstract class to be inherited
by custom entities).
- PhysicState (simple physics component for entities).
- Time (typedefs of std::chrono used all along X-GS).

Modifications in previous existing classes to use / be consistent with
the new ones.

Included some examples:

- Entities/ExampleBallEntity inherits from Entity and implements it.
Initializes, implements logic and updates/render a simple bouncing
ball. It only has two “components”: sf::CircleShape (for the graphics)
and PhysicState for the movement. Bouncing is implemented in the update
method.

- Scenes/ExampleScene and Scenes/ExampleScene2 which implement very
simple scenes (a colored background with a text and an
ExampleBallEntity bouncing around). The first scene changes to the
second after a few seconds, and the second brings back the first again
after the same few seconds. They have also a simple transition system
(fade to/from black) implemented which I should extract to a new
Transition class.

- Scenes/ScenesIdentifiers is a simple enumeration of the avalible
scenes. It is needed by the SceneManager and each scene which ask to
load a new scene.