Bug fix. See description.
Added a mSceneChangeRequest flag for safety. Now, loadScene(Scenes::ID)
only saves the scene id of the scene about to load and changes the flag
to true.
The flag is used to check if there is any scene change request to
prevent updating/rendering/event handling so that it is safe to change
the scene.